// mobile/src/navigation/types.ts
export type AuthStackParamList = {
  Welcome: undefined;
  Login: undefined;
  Register: undefined;
  PublicDoctors: undefined;
  PublicPartners: undefined;
  Contact: undefined;
  About: undefined;
};

export type MainStackParamList = {
  Tabs: undefined;
  Chat: { peerId: string; title: string };
  Profile: undefined;
  Donors: undefined;
  HospitalHistory: undefined;
  Reports: undefined;
  Activity: undefined;
  Messages: undefined;
  Users: undefined;
  RegisterUser: undefined;
  Partners: undefined;
  DoctorsAdmin: undefined;
  Hospitals: undefined;
  Location: undefined;
  NearestDonors: undefined;
};